Fixed some mathematical sections then submitted.
On a last reading I found some bone-headed errors in
the mathematical section --- I left the \ell subscript
on things when it shouldn't have been.  So, I cleaned
that up and fixed it.

Then included referees in the cover letter and fixed a
few typos and submitted it.

This is the version that was submitted.