Cleaned stuff up for SNP simulation
Big note.  I had to add some additional coalescent trees to the simulation because one of them had no suitable SNP ascertained.  This will throw off the seeds from what we had before.  So the results won't be exactly was what we had before.  But that is OK, probably.

I also had to change something in the plateau table section because I used a lower number of Reps than 9 to test it.